Moved createTerm to term folder
