+++ libs/import/src/lib/components/view-extracted-elements/contract/view-extracted-contracts.component.ts
@@ -4,7 +4,8 @@ import { MatTableDataSource } from '@angular/material/table';
 import { MovieService } from '@blockframes/movie/+state';
 import { SheetTab } from '@blockframes/utils/spreadsheet';
 import { Timestamp } from '@blockframes/utils/common-interfaces/timestamp';
-import { createTerm, createMandate, createSale, Mandate, Sale } from '@blockframes/contract/contract/+state/contract.model';
+import { createMandate, createSale, Mandate, Sale } from '@blockframes/contract/contract/+state/contract.model';
+import { createTerm } from '@blockframes/contract/term/+state/term.model';
 import { ContractService } from '@blockframes/contract/contract/+state/contract.service';
 import { Intercom } from 'ng-intercom';
 import { getKeyIfExists } from '@blockframes/utils/helpers';
